server/requirements.txt
aniso8601==1.2.0
appdirs==1.4.2
astroid==1.4.9
click==6.7
facebook-sdk==2.0.0
Flask==0.12
Flask-HTTPAuth==3.2.2
Flask-RESTful==0.3.5
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
lazy-object-proxy==1.2.2
MarkupSafe==0.23
marshmallow==2.13.0
marshmallow-sqlalchemy==0.12.1
mccabe==0.6.1
mysqlclient==1.3.10
nose==1.3.7
packaging==16.8
passlib==1.7.1
pylint==1.6.5
PyMySQL==0.7.10
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
Werkzeug==0.11.15
wrapt==1.10.8
